@@ -13,16 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.16",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "~12.2.16",
-    "@angular/compiler": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.16",
-    "@angular/material": "^12.2.13",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/router": "~12.2.16",
+    "@angular/animations": "^14.2.3",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "~14.2.3",
+    "@angular/compiler": "~14.2.3",
+    "@angular/core": "~14.2.3",
+    "@angular/forms": "~14.2.3",
+    "@angular/material": "^13.3.9",
+    "@angular/platform-browser": "~14.2.3",
+    "@angular/platform-browser-dynamic": "~14.2.3",
+    "@angular/router": "~14.2.3",
     "@fortawesome/angular-fontawesome": "^0.10.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.21",
     "@fortawesome/free-solid-svg-icons": "^5.10.1",
@@ -32,8 +32,8 @@
   },
   "devDependencies": {
     "@angular/cli": "~12.2.18",
-    "@angular/compiler-cli": "~12.2.16",
-    "@angular/language-service": "~12.2.16",
+    "@angular/compiler-cli": "~14.2.3",
+    "@angular/language-service": "~14.2.3",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
@@ -54,6 +54,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.8.3"
   }
 }
